requirements.txt
Pillow
# base
click
git+https://github.com/abey79/vpype#egg=vpype
git+https://github.com/abey79/vpype-text#egg=vpype-text
shapely
numpy
scipy
scikit-image
opencv-python

# dev/test
pytest
black